pyproject.toml
[project]
name = "amplifier-app-blog-creator"
version = "0.1.0"
description = "Complete blog creation workflow: content generation and optional illustration"
authors = [
    {name = "Brian Krabach", email = "briankrabach@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "amplifier-module-style-extraction",
    "amplifier-module-image-generation",
    "amplifier-module-markdown-utils",
    "anthropic>=0.25.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.14",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "markdown>=3.6.0",
    "bleach>=6.0.0",
    "sse-starlette>=2.1.0",
    "starlette>=0.37.0",
    "itsdangerous>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pyright>=1.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
blog-creator = "amplifier_app_blog_creator.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/amplifier_app_blog_creator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv.sources.amplifier-module-style-extraction]
git = "https://github.com/robotdad/amplifier-module-style-extraction"
branch = "main"

[tool.uv.sources.amplifier-module-image-generation]
git = "https://github.com/robotdad/amplifier-module-image-generation"
branch = "main"

[tool.uv.sources.amplifier-module-markdown-utils]
git = "https://github.com/robotdad/amplifier-module-markdown-utils"
branch = "main"

[tool.pyright]
include = ["src", "tests"]
exclude = [".venv", "__pycache__", "*.egg-info"]
pythonVersion = "3.11"
typeCheckingMode = "basic"

[dependency-groups]
dev = [
    "pyright>=1.1.406",
    "ruff>=0.14.1",
]
